backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
psycopg2-binary==2.9.10
pydantic>=2.0.0,<3.0.0
pydantic_core>=2.0.0,<3.0.0
python-dotenv==1.1.1
PyJWT==2.8.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
gunicorn==21.2.0
